requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-client-cache==0.14.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
cattrs==25.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
Flask-WTF==1.2.1
frozenlist==1.8.0
HomeAssistant-API==5.0.2.post2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
platformdirs==4.5.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
requests-cache==1.3.0
simplejson==3.20.2
typing-inspection==0.4.2
typing_extensions==4.15.0
url-normalize==2.2.1
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.4
WTForms==3.1.1
yarl==1.22.0
python-dotenv==1.2.1